+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 - **BREAKING**: `cook_estimator`, `has_gradients`, and `use_named_args` are moved from
   `utils` to `learning`.
 - **BREAKING**: `normalize_dimensions` is moved from `utils` to `space`.
-- **BREAKING**: `branin` has been changed to a `ModelSystem`.
+- **BREAKING**: `branin`, `hart3` has been changed to `ModelSystem`s.
 - Implemented tests for ModelSystems.
-- `hart3`, `hart6`, `poly2`, `peaks` benchmarks systems exist as `ModelSystem`s.
+- `hart6`, `poly2`, `peaks` benchmarks systems exist as `ModelSystem`s.
 - `gold_map` exist as `ModelSystem`.
 - Sampling consolidated. There is now only one sampling function per `Dimension`.
   Different sampling types (at the moment, random value sampling and Latin Hypercube